Cleanup Several `__init__` deficiencies (#59)
This is a grab bag of small stuff. The `__init__` in the main module
exposed a bunch of esoteric and unused stuff. I sorted and titied that.
I also added `__init__.py` instances for the test folders which are
grumpy on my older install for some reason I can't grock having to do
with how pytest handles imports in some version cross section...