Add Tests and Enhance `FakeCamera` consistency with real `Camera`  (#63)
This brings the `FakeCamera` instance up to (tested!) feature parity
with the normal camera instance.

This also includes a pattern to apply tests in to real and fake
instances which is really handy for minimizing test code.

This is the new structure I am looking to transition tests to in this
repo. It has a pattern where a given test can be applied to both the
fake and real camera instances. This means that tests can individually
flag themselves in/out/skipped based on available hardware.

After merging this, the next step is transitioning stuff from the
`tests/old/*` to their new home with pytest runner etc.