Smarter Frame Discard for Testing (#71)
A couple of key changes here. Previously
`scicamera.testing.mature_after_frames_or_timeout()` would return a
`Future` instance. This was often left unchecked, and allowed for actual
errors :grimacing: to creep in despite testing. This diff does several
things:
- The method will now block until the criteria is met, and raise a
standard `TimeoutError` otherwise. This is tested for.
- Several previously broken tests surfaced on this change, so the rest
of the diff is bringing those up to snuff/parity with the new testing
requirements.
- The `FrameRate` control was found to have some inherited oddities
leftover from the "virtual controls" days, so we morphed it into a more
standard getter/setter. This probably deserves a minor version bump.